requirements.txt
flask
python-dotenv
google-generativeai
gradio
langchain
faiss